+++ chalicelib/pvsecret.py
@@ -46,3 +46,35 @@ def get_secrets(stage='prod'):
         secret_json['stripe'] = secret_json['stripe_dev']
 
     return secret_json
+
+
+def get_jwt_signing_key(private=False):
+
+    if not private:
+        secret_name = 'boost-sara/sara-client-public-key'
+    else:
+        secret_name = 'boost-sara/sara-client-private-key'
+
+    region_name = "us-west-2"
+
+    # Create a Secrets Manager client
+    session = boto3.session.Session()
+    client = session.client(
+        service_name='secretsmanager',
+        region_name=region_name
+    )
+
+    try:
+        get_secret_value_response = client.get_secret_value(
+            SecretId=secret_name
+        )
+    except EndpointConnectionError as e:
+        raise e
+    except ClientError as e:
+        # For a list of exceptions thrown, see
+        # https://docs.aws.amazon.com/secretsmanager/latest/apireference/API_GetSecretValue.html
+        raise e
+
+    # Decrypts secret using the associated KMS key.
+    jwt_signing_key = get_secret_value_response['SecretString']
+    return jwt_signing_key